requirements.txt
Flask
python-dotenv
requests
PyMuPDF
pandas
openpyxl
gunicorn